drivers/intel/fsp1_1: Fix vbt_len type
Type of vbt_len should be size_t, change to fix compiler error

Reviewed-on: https://review.coreboot.org/27224
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/drivers/intel/fsp1_1/vbt.c
@@ -26,7 +26,7 @@
 void load_vbt(uint8_t s3_resume, SILICON_INIT_UPD *params)
 {
 	const optionrom_vbt_t *vbt_data = NULL;
-	uint32_t vbt_len;
+	size_t vbt_len;
 
 	/* Check boot mode - for S3 resume path VBT loading is not needed */
 	if (s3_resume) {
